refactor(bridge): remove mechanical analyzer findings (#295)
Repair 18 exact Bridge analyzer findings using behavior-preserving cast, assignment, null-forgiving, and nested-condition changes. Rebuild passes with CS8604 as error; focused bridge/UI suites pass 203/203.

+++ bridge/Commands/ElementCommands.cs
@@ -66,7 +66,7 @@ public static class ElementCommands
             if (rect.Width <= 0 || rect.Height <= 0)
                 return int.MinValue;
 
-            var score = Math.Min((int)(rect.Width * rect.Height / 10_000), 1_000);
+            var score = Math.Min(rect.Width * rect.Height / 10_000, 1_000);
             if (!string.IsNullOrWhiteSpace(SafeString(() => window.Name)))
                 score += 1_000;
             if (!string.IsNullOrWhiteSpace(SafeString(() => window.AutomationId)))
